Reusable loadWtfJson and getRouterPort
